+++ src/components/news-card/NewsCard.css
