+++ src/components/sign-up/SignUp.js
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import PopupWithForm from '../popupWithForm/PopupWithForm';
+import PopupWithForm from '../popup-with-form/PopupWithForm';
 import FormValidator from '../../utils/formValidator';
 
 function SignUp({
